김창권

+++ 01_개발/BodyCare_BackEnd/src/main/java/com/example/bodycare_backend/dao/UserDao.java
@@ -27,5 +27,5 @@ public interface UserDao {
 
     Long insertUser(User user);
 
-    Long deleteUser(String email);
+    int deleteUser(String email);
 }
